requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
duckduckgo-search
python-dotenv
tiktoken
qdrant-client